split model into smaller files
