+++ model_messages.go
@@ -1,0 +1,18 @@
+package main
+
+import (
+	"github.com/charmbracelet/bubbles/list"
+	"github.com/charmbracelet/bubbles/textarea"
+)
+
+type payloadItem struct{ topic, payload string }
+
+func (p payloadItem) FilterValue() string { return p.topic }
+func (p payloadItem) Title() string       { return p.topic }
+func (p payloadItem) Description() string { return p.payload }
+
+type messageState struct {
+	input    textarea.Model
+	payloads []payloadItem
+	list     list.Model // payloadList reused when viewing payloads
+}
