+++ model_topics.go
@@ -5,7 +5,65 @@ import (
 	"sort"
 
 	"github.com/charmbracelet/bubbles/list"
+	"github.com/charmbracelet/bubbles/textinput"
+	"github.com/charmbracelet/bubbles/viewport"
 )
+
+type topicItem struct {
+	title      string
+	subscribed bool
+}
+
+func (t topicItem) FilterValue() string { return t.title }
+func (t topicItem) Title() string       { return t.title }
+func (t topicItem) Description() string {
+	if t.subscribed {
+		return "subscribed"
+	}
+	return "unsubscribed"
+}
+
+type chipBound struct {
+	xPos, yPos    int
+	width, height int
+}
+
+type paneState struct {
+	sel     int
+	page    int
+	m       *model
+	index   int
+	focused bool
+}
+
+func (p *paneState) Focus() {
+	p.focused = true
+	if p.m != nil {
+		p.m.setActivePane(p.index)
+	}
+}
+
+func (p *paneState) Blur() { p.focused = false }
+
+func (p paneState) IsFocused() bool { return p.focused }
+
+func (p paneState) View() string { return "" }
+
+type topicsPanes struct {
+	subscribed   paneState
+	unsubscribed paneState
+	active       int
+}
+
+type topicsState struct {
+	input      textinput.Model
+	items      []topicItem
+	list       list.Model
+	panes      topicsPanes
+	selected   int
+	chipBounds []chipBound
+	vp         viewport.Model
+}
 
 // hasTopic reports whether the given topic already exists in the list.
 func (m *model) hasTopic(topic string) bool {
